+++ lab14/lab14/SawToothGenerator.java
@@ -15,6 +15,6 @@ public class SawToothGenerator implements Generator {
         return normalize(state % period);
     }
     private double normalize(int x) {
-        return x * 2 / period - 1;
+        return (double) x * 2 / period - 1;
     }
 }
